unsigned long long int to unsigned int

+++ inc/array.h
@@ -12,7 +12,7 @@ struct array
 	// expanding which involves copying pointers, and
 	// requesting larger pointer array size than previously we had
 	// this could get expensive and is non productive computation
-	unsigned long long int total_size;
+	unsigned int total_size;
 
 	// we maintain array of pointers each of which point to actual data
 	const void** data_p_p;
@@ -27,14 +27,14 @@ struct array
 
 	// initial_size = the size with which the array was requested from using the initialize_array or get_array function
 	// The array is not allowed to shrink below its initial size
-	unsigned long long int initial_size;
+	unsigned int initial_size;
 };
 
 // returns a new array with total_size as the initial size, with no elements inside,
-array* get_array(unsigned long long int initial_size);
+array* get_array(unsigned int initial_size);
 
 // initializes and gives necessary memory to manage internal element contents
-void initialize_array(array* array_p, unsigned long long int initial_size);
+void initialize_array(array* array_p, unsigned int initial_size);
 
 // frees all the data being held by the array, this function, does not release memory of the actual array structure, only the memory of the components
 // the same array can be reused by calling initialize_array function, after it is deinitialized
@@ -44,22 +44,22 @@ void deinitialize_array(array* array_p);
 void delete_array(array* array_p);
 
 // returns pointer to the data at index = index
-const void* get_element(const array* array_p, unsigned long long int index);
+const void* get_element(const array* array_p, unsigned int index);
 
 // returns 0 for success, -1 for index out of bounds
-int set_element(array* array_p, const void* data_p, unsigned long long int index);
+int set_element(array* array_p, const void* data_p, unsigned int index);
 
 // swap elements of array at given indexes
-void swap_elements(array* array_p, unsigned long long int i1, unsigned long long int i2);
+void swap_elements(array* array_p, unsigned int i1, unsigned int i2);
 
 // perform operation on all the elements of the array, the method operation takes in 2 params, the data_p to operation on and its index
-void for_each_in_array(const array* array_p, void (*operation)(void* data_p, unsigned long long int index, const void* additional_params), const void* additional_params);
+void for_each_in_array(const array* array_p, void (*operation)(void* data_p, unsigned int index, const void* additional_params), const void* additional_params);
 //also ->
-void for_each_non_null_in_array(const array* array_p, void (*operation)(void* data_p, unsigned long long int index, const void* additional_params), const void* additional_params);
+void for_each_non_null_in_array(const array* array_p, void (*operation)(void* data_p, unsigned int index, const void* additional_params), const void* additional_params);
 
 // returns index of the element found in array_p, if compare between the element and data_p returns 0
 // compare > 0, if data_p1 > data_p2 else compare < 0
-unsigned long long int find_first_in_array(const array* array_p, void* data_p, int (*compare)(const void* data_p1, const void* data_p2));
+unsigned int find_first_in_array(const array* array_p, void* data_p, int (*compare)(const void* data_p1, const void* data_p2));
 
 // it expands array, as per the array rules
 void expand_array(array* array_p);
@@ -69,7 +69,7 @@ void expand_array(array* array_p);
 // it returns 1 if the array was shrunk
 // note : The array is not allowed to shrink below its initial size,
 // it shrinks only if the new_total_size is greater than or equal to the initial size
-int shrink_array(array* array_p, unsigned long long int start_index, unsigned long long int end_index);
+int shrink_array(array* array_p, unsigned int start_index, unsigned int end_index);
 
 // prints the array
 void print_array(const array* array_p, void (*print_element)(const void* data_p));
