+++ inc/hashmap.h
@@ -39,10 +39,10 @@ struct hashmap
 	collision_resolution_policy hashmap_policy;
 
 	// if it is using open addressing, this is the node_offset for instructing the linkedlist or binary search tree
-	unsigned long long int node_offset;
+	unsigned int node_offset;
 
 	// hash function to hash the data
-	unsigned long long int (*hash_function)(const void* data);
+	unsigned int (*hash_function)(const void* data);
 
 	// compare data and returns 0 if they are equal, else non-zero
 	// it returns 0 if they are same, >0 if data1 is greater than data2 else it must return <0 value
@@ -52,14 +52,14 @@ struct hashmap
 	void** holder;
 
 	// this is the number of buckets, which are occupied in the hashmap
-	unsigned long long int occupancy;
+	unsigned int occupancy;
 
 	// this is the number of total buckets in the hashmap
-	unsigned long long int total_bucket_count;
+	unsigned int total_bucket_count;
 };
 
 // initializes hashmap and it will depend on initialize_array to give necessary memory to manage internal element contents
-void initialize_hashmap(hashmap* hashmap_p, collision_resolution_policy hashmap_policy, unsigned long long int total_bucket_count, unsigned long long int (*hash_function)(const void* key), int (*compare)(const void* data1, const void* data2), unsigned long long int node_offset);
+void initialize_hashmap(hashmap* hashmap_p, collision_resolution_policy hashmap_policy, unsigned int total_bucket_count, unsigned int (*hash_function)(const void* key), int (*compare)(const void* data1, const void* data2), unsigned int node_offset);
 
 int exists_in_hashmap(hashmap* hashmap_p, const void* data);
